Decompose documentation badge policy
