+++ src/dnadesign/devtools/docs/badges/__init__.py
@@ -1,0 +1,15 @@
+"""Public badge-policy surface for documentation checks."""
+
+from dnadesign.devtools.docs.badges.images import rendered_markdown_images
+from dnadesign.devtools.docs.badges.policy import (
+    ROOT_README_ALLOWED_BADGES,
+    find_markdown_badge_policy_issues,
+    rendered_markdown_badge_lines,
+)
+
+__all__ = [
+    "ROOT_README_ALLOWED_BADGES",
+    "find_markdown_badge_policy_issues",
+    "rendered_markdown_badge_lines",
+    "rendered_markdown_images",
+]
